requirements.txt
absl-py==1.4.0
attrs==22.2.0
bidict==0.22.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.3
Flask-SocketIO==5.3.3
flatbuffers==23.3.3
fonttools==4.39.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.9.1.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
pyalsaaudio==0.9.2
pybase64==1.2.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.4.0
python-socketio==5.8.0
six==1.16.0
Werkzeug==2.2.3
